Add chart mode for sql console result
