+++ src/routes/notebooks.$id.tsx
@@ -17,8 +17,6 @@ import {
 	Loader2,
 	Pencil,
 	Play,
-	Plus,
-	Settings2,
 	Share2,
 	SquareTerminal,
 	Table,
@@ -27,43 +25,22 @@ import {
 	User,
 	X,
 } from "lucide-react";
-import {
-	Children,
-	lazy,
-	Suspense,
-	useCallback,
-	useEffect,
-	useMemo,
-	useRef,
-	useState,
-} from "react";
+import { Children, useCallback, useEffect, useRef, useState } from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
-import type { VisualizationSpec } from "vega-embed";
 import { useAidboxClient } from "../AidboxClient";
 import { ConfirmDialog } from "../components/confirm-dialog";
 import { ResultContent } from "../components/db-console/result-content";
 import { transformToQueryResultItems } from "../components/db-console/tables-view";
 import {
-	ChartIconBar,
-	ChartIconCustom,
-	ChartIconLine,
-	ChartIconPie,
-	ChartIconPyramid,
-} from "../components/notebook-chart/chart-icons";
-import {
-	chartConfigExtensions,
-	foldChartConfigData,
-} from "../components/notebook-chart/fold-data";
+	type ChartConfig,
+	ChartPanel,
+} from "../components/notebook-chart/chart-panel";
 import { NotebookResourcePicker } from "../components/notebook-resource-picker";
 import { HTTP_STATUS_CODES } from "../shared/const";
 import { copyToClipboard } from "../utils/clipboard";
 import { prettyEdn } from "../utils/edn";
 import type { QueryResultItem } from "../webmcp/db-console-context";
-
-const VegaLiteChart = lazy(
-	() => import("../components/notebook-chart/vega-lite-chart"),
-);
 
 type SavedRestResult = {
 	status?: number;
@@ -1108,14 +1085,6 @@ function parseSqlQueryResponse(body: ParamsResp): SqlQueryRunResult {
 	return { ok: true, columns, rows };
 }
 
-type ChartConfig = {
-	type: "bar" | "line" | "pie" | "pyramid" | "custom";
-	x: string;
-	y: string[];
-	color: string;
-	rawSpec?: string;
-};
-
 function parseSqlQueryCellValue(raw: string): {
 	url: string;
 	params: Record<string, string>;
@@ -1184,574 +1153,6 @@ function useRunOnSelect(resourceId: string | undefined, run: () => void) {
 	};
 }
 
-function buildStackedBarSpec(
-	records: Record<string, unknown>[],
-	x: string,
-	measures: string[],
-	color: string,
-): VisualizationSpec {
-	const ys = measures.filter(Boolean);
-	if (ys.length >= 2) {
-		return {
-			$schema: "https://vega.github.io/schema/vega-lite/v6.json",
-			data: { values: records },
-			transform: [{ fold: ys, as: ["measure", "__value"] }],
-			width: "container",
-			height: 280,
-			mark: "bar",
-			encoding: {
-				x: { field: x, type: "nominal" },
-				y: { field: "__value", type: "quantitative", aggregate: "sum" },
-				color: { field: "measure", type: "nominal" },
-			},
-			config: { view: { stroke: null } },
-		} as VisualizationSpec;
-	}
-	const encoding: Record<string, unknown> = {
-		x: { field: x, type: "nominal" },
-		y: { field: ys[0] ?? "", type: "quantitative", aggregate: "sum" },
-	};
-	if (color) {
-		encoding.color = { field: color, type: "nominal" };
-	}
-	return {
-		$schema: "https://vega.github.io/schema/vega-lite/v6.json",
-		data: { values: records },
-		width: "container",
-		height: 280,
-		mark: "bar",
-		encoding,
-		config: { view: { stroke: null } },
-	} as VisualizationSpec;
-}
-
-function buildPyramidSpec(
-	records: Record<string, unknown>[],
-	category: string,
-	measure: string,
-	group: string,
-): VisualizationSpec {
-	if (!group) {
-		return {
-			$schema: "https://vega.github.io/schema/vega-lite/v6.json",
-			data: { values: records },
-			width: "container",
-			height: 280,
-			mark: "bar",
-			encoding: {
-				y: {
-					field: category,
-					type: "nominal",
-					sort: "descending",
-					axis: { title: null },
-				},
-				x: {
-					field: measure,
-					type: "quantitative",
-					aggregate: "sum",
-					title: measure,
-				},
-			},
-			config: { view: { stroke: null }, axis: { grid: false } },
-		} as VisualizationSpec;
-	}
-	const groupValues = Array.from(
-		new Set(records.map((r) => String(r[group]))),
-	).slice(0, 2);
-	const negated = groupValues[1] ?? "";
-	const signed = `datum[${JSON.stringify(group)}] === ${JSON.stringify(negated)} ? -datum[${JSON.stringify(measure)}] : datum[${JSON.stringify(measure)}]`;
-	return {
-		$schema: "https://vega.github.io/schema/vega-lite/v6.json",
-		data: { values: records },
-		transform: [{ calculate: signed, as: "__signed" }],
-		width: "container",
-		height: 280,
-		mark: "bar",
-		encoding: {
-			y: {
-				field: category,
-				type: "nominal",
-				sort: "descending",
-				axis: { title: null },
-			},
-			x: {
-				aggregate: "sum",
-				field: "__signed",
-				type: "quantitative",
-				title: measure,
-				axis: { format: "s" },
-			},
-			color: {
-				field: group,
-				type: "nominal",
-				legend: { orient: "top", title: null },
-			},
-		},
-		config: { view: { stroke: null }, axis: { grid: false } },
-	} as VisualizationSpec;
-}
-
-const CHART_TYPES = [
-	{ value: "bar", label: "Stacked bar", icon: <ChartIconBar /> },
-	{ value: "line", label: "Line", icon: <ChartIconLine /> },
-	{ value: "pie", label: "Pie", icon: <ChartIconPie /> },
-	{ value: "pyramid", label: "Pyramid", icon: <ChartIconPyramid /> },
-	{ value: "custom", label: "Custom config", icon: <ChartIconCustom /> },
-] as const;
-
-function detectFieldType(
-	records: Record<string, unknown>[],
-	field: string,
-): "quantitative" | "temporal" | "nominal" {
-	let numeric = false;
-	for (const r of records) {
-		const v = r[field];
-		if (typeof v === "string" && /^\d{4}-\d{2}/.test(v)) return "temporal";
-		if (typeof v === "number") numeric = true;
-		else if (v != null) return "nominal";
-	}
-	return numeric ? "quantitative" : "nominal";
-}
-
-function buildLineSpec(
-	records: Record<string, unknown>[],
-	x: string,
-	measures: string[],
-	color: string,
-): VisualizationSpec {
-	const ys = measures.filter(Boolean);
-	if (ys.length >= 2) {
-		return {
-			$schema: "https://vega.github.io/schema/vega-lite/v6.json",
-			data: { values: records },
-			transform: [{ fold: ys, as: ["measure", "__value"] }],
-			width: "container",
-			height: 280,
-			mark: { type: "line", point: true },
-			encoding: {
-				x: { field: x, type: detectFieldType(records, x) },
-				y: { field: "__value", type: "quantitative", aggregate: "sum" },
-				color: { field: "measure", type: "nominal" },
-			},
-			config: { view: { stroke: null } },
-		} as VisualizationSpec;
-	}
-	const encoding: Record<string, unknown> = {
-		x: { field: x, type: detectFieldType(records, x) },
-		y: { field: ys[0] ?? "", type: "quantitative", aggregate: "sum" },
-	};
-	if (color) {
-		encoding.color = { field: color, type: "nominal" };
-	}
-	return {
-		$schema: "https://vega.github.io/schema/vega-lite/v6.json",
-		data: { values: records },
-		width: "container",
-		height: 280,
-		mark: { type: "line", point: true },
-		encoding,
-		config: { view: { stroke: null } },
-	} as VisualizationSpec;
-}
-
-function buildPieSpec(
-	records: Record<string, unknown>[],
-	category: string,
-	measure: string,
-): VisualizationSpec {
-	return {
-		$schema: "https://vega.github.io/schema/vega-lite/v6.json",
-		data: { values: records },
-		width: "container",
-		height: 280,
-		mark: { type: "arc", tooltip: true },
-		encoding: {
-			theta: {
-				field: measure,
-				type: "quantitative",
-				aggregate: "sum",
-				stack: true,
-			},
-			color: {
-				field: category,
-				type: "nominal",
-				legend: { orient: "right" },
-			},
-		},
-		config: { view: { stroke: null } },
-	} as VisualizationSpec;
-}
-
-function stripChartData(text: string): string {
-	try {
-		const obj = JSON.parse(text) as Record<string, unknown>;
-		if (obj && typeof obj === "object") delete obj.data;
-		return JSON.stringify(obj, null, 2);
-	} catch {
-		return text;
-	}
-}
-
-function computeChartPreset(
-	columns: string[],
-	records: Record<string, unknown>[],
-): { x: string; y: string; color: string } {
-	const numeric = columns.filter((c) =>
-		records.some((r) => typeof r[c] === "number"),
-	);
-	const x = columns[0] ?? "";
-	const y = numeric.find((c) => c !== x) ?? columns.find((c) => c !== x) ?? x;
-	const color = columns.find((c) => c !== x && c !== y) ?? "";
-	return { x, y, color };
-}
-
-function MeasureList({
-	label,
-	values,
-	columns,
-	onChange,
-}: {
-	label: string;
-	values: string[];
-	columns: string[];
-	onChange: (next: string[]) => void;
-}) {
-	const setAt = (i: number, v: string) =>
-		onChange(values.map((cur, idx) => (idx === i ? v : cur)));
-	const removeAt = (i: number) =>
-		onChange(values.filter((_, idx) => idx !== i));
-	const add = () => {
-		const used = new Set(values);
-		const next = columns.find((c) => !used.has(c)) ?? columns[0] ?? "";
-		onChange([...values, next]);
-	};
-	return (
-		<div className="flex flex-col gap-1.5">
-			<span className="typo-label-tiny uppercase tracking-wide text-text-tertiary">
-				{label}
-			</span>
-			{values.map((v, i) => (
-				<div key={v} className="flex items-center gap-1">
-					<HSComp.Select value={v} onValueChange={(nv) => setAt(i, nv)}>
-						<HSComp.SelectTrigger className="w-full h-8">
-							<HSComp.SelectValue placeholder="—" />
-						</HSComp.SelectTrigger>
-						<HSComp.SelectContent>
-							{columns.map((c) => (
-								<HSComp.SelectItem key={c} value={c}>
-									{c}
-								</HSComp.SelectItem>
-							))}
-						</HSComp.SelectContent>
-					</HSComp.Select>
-					{values.length > 1 && (
-						<HSComp.IconButton
-							icon={<X />}
-							aria-label="Remove measure"
-							onClick={() => removeAt(i)}
-						/>
-					)}
-				</div>
-			))}
-			{columns.length > values.length && (
-				<HSComp.Button
-					variant="ghost"
-					size="small"
-					onClick={add}
-					className="self-start"
-				>
-					<Plus />
-					Add measure
-				</HSComp.Button>
-			)}
-		</div>
-	);
-}
-
-function ColumnSelect({
-	label,
-	value,
-	columns,
-	onChange,
-	allowNone,
-}: {
-	label: string;
-	value: string;
-	columns: string[];
-	onChange: (v: string) => void;
-	allowNone?: boolean;
-}) {
-	return (
-		<div className="flex flex-col gap-1.5">
-			<span className="typo-label-tiny uppercase tracking-wide text-text-tertiary">
-				{label}
-			</span>
-			<HSComp.Select
-				value={value === "" ? "__none__" : value}
-				onValueChange={(v) => onChange(v === "__none__" ? "" : v)}
-			>
-				<HSComp.SelectTrigger className="w-full h-8">
-					<HSComp.SelectValue placeholder="—" />
-				</HSComp.SelectTrigger>
-				<HSComp.SelectContent>
-					{allowNone && (
-						<HSComp.SelectItem value="__none__">None</HSComp.SelectItem>
-					)}
-					{columns.map((c) => (
-						<HSComp.SelectItem key={c} value={c}>
-							{c}
-						</HSComp.SelectItem>
-					))}
-				</HSComp.SelectContent>
-			</HSComp.Select>
-		</div>
-	);
-}
-
-function NotebookChartPanel({
-	columns,
-	rows,
-	chart,
-	onChartChange,
-	editable,
-}: {
-	columns: string[];
-	rows: unknown[][];
-	chart?: ChartConfig;
-	onChartChange?: (config: ChartConfig) => void;
-	editable: boolean;
-}) {
-	const [showSettings, setShowSettings] = useState(editable);
-	const [chartType, setChartType] = useState<
-		"bar" | "line" | "pie" | "pyramid" | "custom"
-	>(chart?.type ?? "bar");
-	const [rawSpec, setRawSpec] = useState(chart?.rawSpec ?? "");
-	const [chartError, setChartError] = useState<string | null>(null);
-
-	const records = useMemo(() => {
-		return rows.map((row) => {
-			const obj: Record<string, unknown> = {};
-			columns.forEach((col, i) => {
-				obj[col] = row[i];
-			});
-			return obj;
-		});
-	}, [columns, rows]);
-
-	const preset = useMemo(
-		() => computeChartPreset(columns, records),
-		[columns, records],
-	);
-	const [xField, setXField] = useState(chart?.x ?? preset.x);
-	const [yFields, setYFields] = useState<string[]>(chart?.y ?? [preset.y]);
-	const [colorField, setColorField] = useState(chart?.color ?? preset.color);
-
-	const colsKey = JSON.stringify(columns);
-	const [prevColsKey, setPrevColsKey] = useState(colsKey);
-	if (prevColsKey !== colsKey) {
-		setPrevColsKey(colsKey);
-		setXField(preset.x);
-		setYFields([preset.y]);
-		setColorField(preset.color);
-	}
-
-	const reportRef = useRef(onChartChange);
-	reportRef.current = onChartChange;
-	const firstReport = useRef(true);
-	useEffect(() => {
-		if (firstReport.current) {
-			firstReport.current = false;
-			return;
-		}
-		reportRef.current?.({
-			type: chartType,
-			x: xField,
-			y: yFields,
-			color: colorField,
-			rawSpec:
-				chartType === "custom" && rawSpec ? stripChartData(rawSpec) : undefined,
-		});
-	}, [chartType, xField, yFields, colorField, rawSpec]);
-
-	const formSpec = useMemo(() => {
-		const firstY = yFields[0] ?? "";
-		if (chartType === "pyramid")
-			return buildPyramidSpec(records, xField, firstY, colorField);
-		if (chartType === "line")
-			return buildLineSpec(records, xField, yFields, colorField);
-		if (chartType === "pie")
-			return buildPieSpec(records, colorField || xField, firstY);
-		return buildStackedBarSpec(records, xField, yFields, colorField);
-	}, [chartType, records, xField, yFields, colorField]);
-
-	const multiMeasure =
-		(chartType === "bar" || chartType === "line") && yFields.length >= 2;
-
-	const { spec, error: specError } = useMemo<{
-		spec: VisualizationSpec | null;
-		error: string | null;
-	}>(() => {
-		if (chartType !== "custom") return { spec: formSpec, error: null };
-		try {
-			const parsed = JSON.parse(rawSpec) as Record<string, unknown>;
-			if (parsed && typeof parsed === "object" && !("data" in parsed)) {
-				parsed.data = { values: records };
-			}
-			return { spec: parsed as VisualizationSpec, error: null };
-		} catch (e) {
-			return { spec: null, error: e instanceof Error ? e.message : String(e) };
-		}
-	}, [chartType, rawSpec, formSpec, records]);
-
-	const [prevSpec, setPrevSpec] = useState(spec);
-	if (prevSpec !== spec) {
-		setPrevSpec(spec);
-		setChartError(null);
-	}
-
-	const chartBody = (
-		<>
-			{spec ? (
-				<Suspense
-					fallback={
-						<div className="flex items-center justify-center h-[280px] text-text-tertiary">
-							<Loader2 className="size-4 animate-spin" />
-						</div>
-					}
-				>
-					<VegaLiteChart
-						spec={spec}
-						onError={(e) => setChartError(String(e))}
-					/>
-				</Suspense>
-			) : (
-				<div className="px-1 py-2 typo-body-xs text-critical-default">
-					Invalid Vega-Lite config: {specError}
-				</div>
-			)}
-			{spec && chartError && (
-				<div className="px-1 py-2 typo-body-xs text-critical-default">
-					{chartError}
-				</div>
-			)}
-		</>
-	);
-
-	if (!showSettings) {
-		return (
-			<div className="flex flex-col max-h-[460px]">
-				<div className="flex justify-end px-2 pt-2">
-					<HSComp.Toggle
-						variant="outline"
-						pressed={false}
-						onPressedChange={() => setShowSettings(true)}
-						aria-label="Chart settings"
-					>
-						<Settings2 />
-						Settings
-					</HSComp.Toggle>
-				</div>
-				<div className="min-h-0 overflow-auto p-3 pt-2">{chartBody}</div>
-			</div>
-		);
-	}
-
-	return (
-		<div className="h-[400px]">
-			<HSComp.ResizablePanelGroup direction="horizontal">
-				<HSComp.ResizablePanel defaultSize={60} minSize={30}>
-					<div className="h-full overflow-auto p-3">{chartBody}</div>
-				</HSComp.ResizablePanel>
-				<HSComp.ResizableHandle withHandle />
-				<HSComp.ResizablePanel defaultSize={40} minSize={20}>
-					<div className="flex h-full flex-col bg-bg-primary">
-						<div className="flex items-center justify-between pl-3 pr-2 h-9 border-b border-border-default shrink-0">
-							<div className="flex items-center gap-1">
-								{CHART_TYPES.map((t) => (
-									<HSComp.Tooltip key={t.value}>
-										<HSComp.TooltipTrigger asChild>
-											<HSComp.IconButton
-												icon={t.icon}
-												aria-label={t.label}
-												onClick={() => {
-													if (t.value === "custom" && !rawSpec) {
-														setRawSpec(JSON.stringify(formSpec, null, 2));
-													}
-													setChartType(t.value);
-												}}
-												className={
-													chartType === t.value
-														? "border border-border-link text-text-info-primary bg-bg-tertiary"
-														: ""
-												}
-											/>
-										</HSComp.TooltipTrigger>
-										<HSComp.TooltipContent side="bottom">
-											{t.label}
-										</HSComp.TooltipContent>
-									</HSComp.Tooltip>
-								))}
-							</div>
-							<HSComp.IconButton
-								icon={<X />}
-								aria-label="Close chart settings"
-								onClick={() => setShowSettings(false)}
-							/>
-						</div>
-						{chartType === "custom" ? (
-							<div className="flex-1 min-h-0 overflow-auto">
-								<HSComp.CodeEditor
-									mode="json"
-									defaultValue={rawSpec}
-									onChange={setRawSpec}
-									lineNumbers={false}
-									additionalExtensions={chartConfigExtensions}
-									viewCallback={foldChartConfigData}
-								/>
-							</div>
-						) : (
-							<div className="flex flex-col gap-3 p-3 min-h-0 flex-1 overflow-auto">
-								{chartType !== "pie" && (
-									<ColumnSelect
-										label="X axis"
-										value={xField}
-										columns={columns}
-										onChange={setXField}
-									/>
-								)}
-								{chartType === "bar" || chartType === "line" ? (
-									<MeasureList
-										label="Y axis"
-										values={yFields}
-										columns={columns}
-										onChange={setYFields}
-									/>
-								) : (
-									<ColumnSelect
-										label={chartType === "pie" ? "Value" : "Y axis"}
-										value={yFields[0] ?? ""}
-										columns={columns}
-										onChange={(v) => setYFields([v])}
-									/>
-								)}
-								{!multiMeasure && (
-									<ColumnSelect
-										label={chartType === "pie" ? "Category" : "Color"}
-										value={colorField}
-										columns={columns}
-										onChange={setColorField}
-										allowNone={chartType !== "pie"}
-									/>
-								)}
-							</div>
-						)}
-					</div>
-				</HSComp.ResizablePanel>
-			</HSComp.ResizablePanelGroup>
-		</div>
-	);
-}
-
 function SqlQueryResultView({
 	result,
 	canChart,
@@ -1792,7 +1193,7 @@ function SqlQueryResultView({
 					{result.error}
 				</pre>
 			) : canChart && viewMode === "chart" ? (
-				<NotebookChartPanel
+				<ChartPanel
 					columns={result.columns}
 					rows={result.rows}
 					chart={chart}
@@ -1872,7 +1273,7 @@ export function SqlQueryCellView({
 		parsed.view ?? "table",
 	);
 
-	const canChart = isView && !!result && result.ok && result.rows.length > 0;
+	const canChart = !!result && result.ok && result.rows.length > 0;
 
 	const run = async () => {
 		if (!lib?.id) return;
